scripts/requirements.txt
blinker==1.8.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dissect.cstruct==3.15.dev1
dissect.esedb==3.13
dissect.util==3.17.dev1
dnspython==2.6.1
dsinternals==1.2.4
Flask==3.0.3
future==1.0.0
impacket==0.11.0
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.5
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pyOpenSSL==24.1.0
six==1.16.0
termcolor==2.4.0
Werkzeug==3.0.3
zipp==3.19.2
